pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Abi Raja <abimanyuraja@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = "^0.25.0"
python-dotenv = "^1.0.0"
databases = {extras = ["aiomysql"], version = "^0.8.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pyright = "^1.1.345"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
